// src/lib/types.ts
export type WeeklyData = {
  Date: Date;
  Value: number; // Target Column (Value)
  Orders: number; // Exogenous Column (Orders)
  Forecast?: number; // Forecast Data (if available)
  ForecastLower?: number; // Lower bound of forecast
  ForecastUpper?: number; // Upper bound of forecast
  CreatedDate: Date; // Created Date
};